new feature; layer list filter
Now in the window layer list you are able to filter the listed layers,
by input a reg-expression into the lineEdit object. Every changing
inside the lineEdit will renew the layer list.

The buttons hide and show all layers now have only affect to the shown
layers in the list box.

Known bugs:

At the moment editing, removing and adding a layer are not able to renew
the layer list box.


Workaround:

After using one of this option you have to change the reg-expression for
a renew.


Planed issue:

To solve this problem I plan to add the QG_LayerModel class
to the RS_LayerList class.